Finished header, buttons, almost done nav
