+++ src/lib/Navigation.svelte
@@ -1,34 +1,16 @@
 <!-- add mobile friendly hamburger.  -->
-<nav class="navbar" role="navigation" aria-label="main navigation">
+<nav class="navbar">
   <div class="navbar-brand">
-    <a class="navbar-item" href="/home">
+    <a class="navbar-item" href="/">
       <b>Flatter</b>
     </a>
-    <a role="button" class="navbar-burger" aria-label="menu" aria-expanded="false" data-target="navbarBasicExample" id="myNavbarBurger">
-      <span aria-hidden="true"></span>
-      <span aria-hidden="true"></span>
-      <span aria-hidden="true"></span>
-    </a>
-  </div>
-  <div id="navbarBasicExample" class="navbar-menu">
-    <!-- Add your navigation links here -->
-    <div class="navbar-start">
-      <!-- Navbar items on the left -->
-      <a href="/home" class="navbar-item">Home</a>
-      <a href="/" class="navbar-item">Calculator</a>
-      <!-- <a class="navbar-item">More</a> -->
-    </div>
-    <div class="navbar-end"></div>
-    <div class="navbar-item has-dropdown is-hoverable">
-      <a class="navbar-link">More</a>
-      <div class="navbar-dropdown">
-        <a href="/about" class="navbar-item">About</a>
-        <a href="/contact" class="navbar-item">Contact</a>
-      </div>
+    <div id="navbarBasicExample" class="navbar-menu">
+      <!-- Add your navigation links here -->
 
-      <!-- Navbar items on the right -->
-      <!-- <a class="navbar-item">Sign up</a>
-      <a class="navbar-item">Log in</a> -->
+      <!-- Navbar items on the left -->
+      <a href="/" class="navbar-item">Home</a>
+      <a href="/calculator" class="navbar-item">Calculator</a>
+      <!-- <a class="navbar-item">More</a> -->
     </div>
   </div>
 </nav>
